Move private channel and short-conf tests to new module
