+++ lightning/src/ln/mod.rs
@@ -54,6 +54,9 @@ mod functional_tests;
 mod payment_tests;
 #[cfg(test)]
 #[allow(unused_mut)]
+mod priv_short_conf_tests;
+#[cfg(test)]
+#[allow(unused_mut)]
 mod chanmon_update_fail_tests;
 #[cfg(test)]
 #[allow(unused_mut)]
